Fix EM to three components
